r move innerclass to package.
