+++ approvaltests/src/main/java/org/approvaltests/reporters/intellij/IntelliJPathResolver.java
@@ -31,13 +31,13 @@ public class IntelliJPathResolver {
         return Files.walk(Paths.get(channelsPath), 1, FileVisitOption.FOLLOW_LINKS)
                 .map(Path::getFileName)
                 .map(Objects::toString)
-                .filter(IntelliJCommunityReporter.Version::isVersionFile)
-                .map(IntelliJCommunityReporter.Version::new)
+                .filter(Version::isVersionFile)
+                .map(Version::new)
                 .max(Comparator.naturalOrder())
                 .map(this::getPath);
     }
 
-    private Path getPath(IntelliJCommunityReporter.Version version) {
+    private Path getPath(Version version) {
         String runtimeSuffix = "/bin/idea64.exe";
         return Paths.get(channelsPath + version.version + runtimeSuffix).toAbsolutePath();
     }
